LayoutTests/inspector-obsolete/timeline/timeline-script-tag-2-expected.txt

CONSOLE MESSAGE: line 1: SCRIPT TAG
Tests the Timeline API instrumentation of a script tag with an external script.

EvaluateScript Properties:
{
    children : <object>
    counters : <object>
    data : {
        lineNumber : <number>
        url : <string>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "EvaluateScript"
    usedHeapSize : <number>
}
